README.md: face-detection: Crop face before running face landmark model
The Face Detection mode runs inference using the face detection model
and uses the crop containing the identified face as an input for the
face landmark model. The user no longer is needed to be 15cm away from
the camera to identify the face accurately, so we should remove the Face
Detection note.

Signed-off-by: Ayman Chaudhry <ayman.chaudhry.kc@renesas.com>