linux-renesas: Add build time option to disable OV5645 auto focus
Sometimes it is not desirable to have the auto focus enabled.
At the moment there is no V4L2 support to enable/disable OV5645 auto focus
support on the fly, so add a build time option to add/remove the relevant
code from the driver.

Not an ideal solution, but a quick one.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>